refactor: extract engine/ module from lib.rs (#112 PR 3)
Move hook pipeline, shim execution, exec command, and AI guard
from lib.rs into src/engine/ directory:

- engine/hook.rs (725L): Hook pipeline — input parsing, command
  checking, protected file detection, cursor-hook, fuzz entry points.
  DO NOT SPLIT (security-critical, T8 DREAD 9.0 fail-close).
- engine/shim.rs (577L): PATH shim execution, run_command pipeline,
  hook integrity checking, try_audit_append, update_baseline_silent.
- engine/exec.rs (34L): `omamori exec` subcommand handler.
- engine/guard.rs (25L): AI environment guard (T3, 7 call sites).

lib.rs: 2,893 → 1,534 lines (-47%).

Tests moved: 21 (13 hook + 8 shim). Test count unchanged at 473.
Fuzz entry points re-exported via pub use in lib.rs.
