+++ src/lib.rs
@@ -3,6 +3,7 @@ pub mod audit;
 pub mod config;
 pub mod context;
 pub mod detector;
+mod engine;
 pub mod installer;
 pub mod integrity;
 pub mod rules;
@@ -13,17 +14,23 @@ use std::env;
 use std::ffi::OsString;
 use std::path::{Path, PathBuf};
 
-use util::{
-    binary_name, clone_lossy, parse_config_flag, print_usage, resolve_real_command,
-    should_block_for_sudo, usage_text,
-};
+// --- engine submodule imports ---
+use engine::exec::run_exec_command;
+use engine::guard::guard_ai_config_modification;
+use engine::hook::{run_cursor_hook, run_hook_check};
+use engine::shim::{emit_config_warnings, run_shim, update_baseline_silent};
 
-use actions::{ActionExecutor, ActionOutcome, SystemOps};
-use audit::AuditLogger;
+// --- util imports ---
+use util::{binary_name, parse_config_flag, print_usage, usage_text};
+
+// --- crate module imports ---
 use config::{ConfigLoadResult, load_config};
 use detector::evaluate_detectors;
 use installer::{InstallOptions, default_base_dir, install, uninstall};
 use rules::{CommandInvocation, RuleConfig, match_rule};
+
+// Re-export fuzz entry points (pub API for fuzz harness)
+pub use engine::hook::{fuzz_check_command_for_hook, fuzz_extract_hook_input};
 
 #[derive(Debug)]
 pub enum AppError {
@@ -245,32 +252,6 @@ fn run_policy_test_command(args: &[OsString]) -> Result<i32, AppError> {
     if failures == 0 { Ok(0) } else { Ok(1) }
 }
 
-fn run_exec_command(args: &[OsString]) -> Result<i32, AppError> {
-    let mut position = 2usize;
-    let config_path = if args.get(position).and_then(|item| item.to_str()) == Some("--config") {
-        let value = args
-            .get(position + 1)
-            .ok_or_else(|| AppError::Usage("--config requires a path".to_string()))?;
-        position += 2;
-        Some(PathBuf::from(value))
-    } else {
-        None
-    };
-
-    if args.get(position).and_then(|item| item.to_str()) != Some("--") {
-        return Err(AppError::Usage(format!(
-            "exec requires `--` before the target command\n\n{}",
-            usage_text()
-        )));
-    }
-
-    let program = args
-        .get(position + 1)
-        .ok_or_else(|| AppError::Usage("missing command after `--`".to_string()))?;
-    let command_args = &args[(position + 2)..];
-    run_command(binary_name(program), command_args, config_path.as_deref())
-}
-
 fn run_install_command(args: &[OsString]) -> Result<i32, AppError> {
     let mut base_dir = default_base_dir();
     let mut source_exe = installer::resolve_stable_exe_path(&env::current_exe()?);
@@ -471,358 +452,6 @@ fn run_uninstall_command(args: &[OsString]) -> Result<i32, AppError> {
     );
     println!("Removed {} file(s)", result.removed_entries.len());
     Ok(0)
-}
-
-fn run_shim(program: &str, args: &[OsString]) -> Result<i32, AppError> {
-    let base_dir = default_base_dir();
-
-    // Step 1: Lightweight integrity canary (stat + readlink, ~0.05ms)
-    if let Some(warning) = integrity::canary(&base_dir, program) {
-        eprintln!("omamori[health]: {warning}");
-    }
-
-    // Step 1b: v0.4 → v0.5 migration — create baseline if missing
-    if !integrity::baseline_path(&base_dir).exists() && base_dir.join("shim").exists() {
-        update_baseline_silent(&base_dir);
-        eprintln!(
-            "omamori[health]: integrity baseline created. Run `omamori status` for full check."
-        );
-    }
-
-    // Step 2: Hook version + content hash check, regenerate if needed
-    let hooks_regenerated = ensure_hooks_current();
-
-    // Step 2b: Auto-setup Codex hooks if CODEX_CI detected but not configured
-    let codex_setup = installer::auto_setup_codex_if_needed(&base_dir);
-
-    // Step 3: If hooks were regenerated or Codex was set up, update baseline
-    if hooks_regenerated || codex_setup {
-        update_baseline_silent(&base_dir);
-    }
-
-    // Step 4: Run the actual command
-    run_command(program.to_string(), args, None)
-}
-
-/// Check if hooks are current; if not, regenerate them.
-/// Runs at shim startup. Failures are non-fatal (warn only).
-/// Returns `true` if hooks were regenerated.
-///
-/// Two-level check:
-/// 1. Version mismatch → regenerate (existing behavior, e.g. after upgrade)
-/// 2. Version match but content hash mismatch → regenerate (T2 attack: AI keeps
-///    version comment but rewrites hook body, e.g. `exit 2` → `exit 0`)
-fn ensure_hooks_current() -> bool {
-    ensure_hooks_current_at(&default_base_dir())
-}
-
-/// Testable version of `ensure_hooks_current` that accepts a base directory.
-///
-/// Two-level check:
-/// 1. Version mismatch → regenerate (existing behavior, e.g. after upgrade)
-/// 2. Version match but content hash mismatch → regenerate (T2 attack: AI keeps
-///    version comment but rewrites hook body, e.g. `exit 2` → `exit 0`)
-fn ensure_hooks_current_at(base_dir: &Path) -> bool {
-    let hook_path = base_dir.join("hooks/claude-pretooluse.sh");
-
-    let content = match std::fs::read_to_string(&hook_path) {
-        Ok(c) => c,
-        Err(_) => return false, // No hooks file = not installed via install --hooks, skip
-    };
-
-    let hook_version = installer::parse_hook_version(&content);
-    let version_matches = hook_version == Some(env!("CARGO_PKG_VERSION"));
-
-    if !version_matches {
-        // Level 1: version mismatch → regenerate
-        let current = hook_version
-            .map(|v| v.to_string())
-            .unwrap_or_else(|| "unknown".to_string());
-
-        match installer::regenerate_hooks(base_dir) {
-            Ok(()) => {
-                eprintln!(
-                    "omamori: hooks updated ({} → {})",
-                    current,
-                    env!("CARGO_PKG_VERSION")
-                );
-                return true;
-            }
-            Err(e) => {
-                eprintln!(
-                    "omamori: failed to update hooks ({}). Run: omamori install --hooks",
-                    e
-                );
-            }
-        }
-        return false;
-    }
-
-    // Level 2: version matches → check content hash (T2 attack detection)
-    let expected = installer::render_hook_script();
-    let expected_hash = installer::hook_content_hash(&expected);
-    let actual_hash = installer::hook_content_hash(&content);
-
-    if expected_hash != actual_hash {
-        match installer::regenerate_hooks(base_dir) {
-            Ok(()) => {
-                eprintln!("omamori: hooks content mismatch detected — regenerated");
-                return true;
-            }
-            Err(e) => {
-                eprintln!(
-                    "omamori: failed to regenerate hooks ({}). Run: omamori install --hooks",
-                    e
-                );
-            }
-        }
-    }
-
-    false
-}
-
-/// Attempt to append an audit event. On failure:
-/// - Always emits a WARNING to stderr
-/// - In strict mode, returns `Some(1)` to signal the caller should exit
-/// - In non-strict mode, returns `None` (command execution is not affected)
-fn try_audit_append(
-    logger: &audit::AuditLogger,
-    event: audit::AuditEvent,
-    strict: bool,
-) -> Option<i32> {
-    if let Err(e) = logger.append(event) {
-        eprintln!("omamori warning: audit log write failed: {e}");
-        eprintln!("  Command execution was not affected — this is a logging issue only.");
-        eprintln!(
-            "  To fix: check permissions on ~/.local/share/omamori/ or run omamori install --hooks"
-        );
-        if strict {
-            eprintln!("omamori error: audit strict mode — blocking because audit log is required");
-            return Some(1);
-        }
-    }
-    None
-}
-
-/// Silently update integrity baseline. Used after hook regen or config changes.
-/// Failures are non-fatal (warn to stderr).
-fn update_baseline_silent(base_dir: &Path) {
-    match integrity::generate_baseline(base_dir) {
-        Ok(baseline) => {
-            if let Err(e) = integrity::write_baseline(base_dir, &baseline) {
-                eprintln!("omamori[health]: failed to update baseline: {e}");
-            }
-        }
-        Err(e) => {
-            eprintln!("omamori[health]: failed to generate baseline: {e}");
-        }
-    }
-}
-
-fn run_command(
-    program: String,
-    args: &[OsString],
-    config_path: Option<&Path>,
-) -> Result<i32, AppError> {
-    let load_result = load_config(config_path)?;
-    emit_config_warnings(&load_result);
-
-    let invocation =
-        CommandInvocation::new(program.clone(), args.iter().map(clone_lossy).collect());
-    let env_pairs = env::vars().collect::<Vec<_>>();
-    let detection = evaluate_detectors(&load_result.config.detectors, &env_pairs);
-
-    // Sudo check: always evaluated, regardless of AI detection.
-    // Root-privilege destructive operations must be blocked even in non-AI environments.
-    if should_block_for_sudo() {
-        let outcome = ActionOutcome::Blocked {
-            message:
-                "omamori blocked this command because it was invoked via sudo/elevated privileges"
-                    .to_string(),
-        };
-        eprintln!("{}", outcome.message());
-        for warning in &detection.warnings {
-            eprintln!("omamori warning: {warning}");
-        }
-        if let Some(logger) = AuditLogger::from_config(&load_result.config.audit) {
-            let event =
-                logger.create_event(&invocation, None, &detection.matched_detectors, &outcome);
-            if let Some(code) = try_audit_append(&logger, event, load_result.config.audit.strict) {
-                return Ok(code);
-            }
-        }
-        return Ok(outcome.exit_code());
-    }
-
-    // Non-protected fast path: no AI environment detected = human terminal.
-    // This is NOT a security boundary — omamori only guards AI-initiated commands.
-    // match_rule, context evaluation, and ActionExecutor are not constructed.
-    if !detection.protected {
-        let resolved = resolve_real_command(&program)?;
-        let status = std::process::Command::new(&resolved)
-            .args(&invocation.args)
-            .status()?;
-        let exit_code = actions::exit_code_from_status(status);
-        let outcome = ActionOutcome::PassedThrough { exit_code };
-        for warning in &detection.warnings {
-            eprintln!("omamori warning: {warning}");
-        }
-        if let Some(logger) = AuditLogger::from_config(&load_result.config.audit) {
-            let event =
-                logger.create_event(&invocation, None, &detection.matched_detectors, &outcome);
-            if let Some(code) = try_audit_append(&logger, event, load_result.config.audit.strict) {
-                return Ok(code);
-            }
-        }
-        return Ok(exit_code);
-    }
-
-    // --- Protected path: AI environment detected. Full evaluation. ---
-
-    // Strict mode: block if audit HMAC secret is unavailable (opt-in).
-    // This prevents AI commands from running without tamper-evident logging.
-    if load_result.config.audit.strict && load_result.config.audit.enabled {
-        match AuditLogger::from_config(&load_result.config.audit) {
-            Some(logger) if !logger.secret_available() => {
-                eprintln!("omamori: audit strict mode — HMAC secret unavailable, blocking command");
-                eprintln!(
-                    "omamori: to fix, re-create the secret or set audit.strict = false in config.toml"
-                );
-                return Ok(1);
-            }
-            None => {
-                eprintln!(
-                    "omamori: audit strict mode — audit logger unavailable, blocking command"
-                );
-                return Ok(1);
-            }
-            _ => {} // secret available, proceed normally
-        }
-    }
-
-    let matched_rule = match_rule(&load_result.config.rules, &invocation);
-    let detector_env_keys: Vec<String> = load_result
-        .config
-        .detectors
-        .iter()
-        .map(|d| d.env_key.clone())
-        .filter(|k| !k.is_empty() && !k.contains('='))
-        .collect();
-
-    // Context-aware evaluation: compute effective rule (may differ from matched_rule)
-    let context_override: Option<RuleConfig> = if let (Some(rule), Some(ctx_config)) =
-        (matched_rule, &load_result.config.context)
-    {
-        // Tier 1: path-based evaluation
-        let ctx = context::evaluate_context(&invocation, rule, ctx_config);
-        let tier1_override = if let Some(override_action) = ctx.action_override {
-            eprintln!(
-                "omamori: {} {} → {} ({}, original: {})",
-                invocation.program,
-                invocation.target_args().join(" "),
-                override_action.as_str(),
-                ctx.reason,
-                rule.action.as_str(),
-            );
-            let mut overridden = rule.clone();
-            overridden.message = Some(override_action.context_message(&ctx.reason));
-            overridden.action = override_action;
-            Some(overridden)
-        } else {
-            if !ctx.reason.contains("no target paths") && !ctx.reason.contains("no context pattern")
-            {
-                eprintln!("omamori warning: {}", ctx.reason);
-            }
-            None
-        };
-
-        // Tier 2: git-aware evaluation (skip if Tier 1 already escalated to Block)
-        let is_escalated = tier1_override
-            .as_ref()
-            .is_some_and(|r| matches!(r.action, rules::ActionKind::Block));
-
-        if !is_escalated {
-            if let Some(git_ctx) =
-                context::evaluate_git_context(&invocation, &ctx_config.git, &detector_env_keys)
-            {
-                if let Some(git_action) = git_ctx.action_override {
-                    eprintln!(
-                        "omamori: {} {} → {} ({}, original: {})",
-                        invocation.program,
-                        invocation.args.join(" "),
-                        git_action.as_str(),
-                        git_ctx.reason,
-                        rule.action.as_str(),
-                    );
-                    let mut overridden = rule.clone();
-                    overridden.message = Some(git_action.context_message(&git_ctx.reason));
-                    overridden.action = git_action;
-                    Some(overridden)
-                } else {
-                    if !git_ctx.reason.contains("skipping") {
-                        eprintln!("omamori: {}", git_ctx.reason);
-                    }
-                    tier1_override
-                }
-            } else {
-                tier1_override
-            }
-        } else {
-            tier1_override
-        }
-    } else {
-        None
-    };
-
-    let effective_rule = match (&context_override, matched_rule) {
-        (Some(overridden), _) => Some(overridden),
-        (None, Some(rule)) => Some(rule),
-        _ => None,
-    };
-
-    let resolved_program = resolve_real_command(&program)?;
-    let mut executor =
-        ActionExecutor::new(SystemOps::new(resolved_program, detector_env_keys.clone()));
-
-    let outcome = if let Some(rule) = effective_rule {
-        let outcome = executor.execute(&invocation, rule)?;
-        match &outcome {
-            ActionOutcome::Blocked { .. } | ActionOutcome::Failed { .. } => {
-                eprintln!("{}", outcome.message());
-            }
-            ActionOutcome::Trashed { message, .. } | ActionOutcome::MovedTo { message, .. } => {
-                eprintln!("{message}");
-            }
-            _ => {}
-        }
-        outcome
-    } else {
-        executor.exec_passthrough(&invocation)?
-    };
-
-    for warning in &detection.warnings {
-        eprintln!("omamori warning: {warning}");
-    }
-
-    if let Some(logger) = AuditLogger::from_config(&load_result.config.audit) {
-        let event = logger.create_event(
-            &invocation,
-            effective_rule,
-            &detection.matched_detectors,
-            &outcome,
-        );
-        if let Some(code) = try_audit_append(&logger, event, load_result.config.audit.strict) {
-            return Ok(code);
-        }
-    }
-
-    Ok(outcome.exit_code())
-}
-
-fn emit_config_warnings(load_result: &ConfigLoadResult) {
-    for warning in &load_result.warnings {
-        eprintln!("omamori warning: {warning}");
-    }
 }
 
 fn run_status_command(args: &[OsString]) -> Result<i32, AppError> {
@@ -1160,538 +789,6 @@ fn audit_usage() -> &'static str {
   omamori audit key rotate                       Rotate HMAC signing key"
 }
 
-/// Cursor `beforeShellExecution` hook handler.
-/// Reads JSON from stdin, checks command via shared hook pipeline,
-/// writes JSON response to stdout. All logs go to stderr only.
-fn run_cursor_hook() -> Result<i32, AppError> {
-    use std::io::Read;
-
-    let mut input = String::new();
-    std::io::stdin().read_to_string(&mut input)?;
-
-    let command = match serde_json::from_str::<serde_json::Value>(&input) {
-        Ok(v) => match v.get("command") {
-            Some(c) if c.is_string() => c.as_str().unwrap().to_string(),
-            Some(_) | None => {
-                // command key missing or not a string → malformed protocol → deny
-                eprintln!("omamori cursor-hook: missing or invalid 'command' field");
-                print_cursor_response(false, "deny", Some("omamori: malformed hook input"), None);
-                return Ok(0);
-            }
-        },
-        Err(_) => {
-            eprintln!("omamori cursor-hook: failed to parse stdin JSON");
-            print_cursor_response(false, "deny", Some("omamori: malformed hook input"), None);
-            return Ok(0);
-        }
-    };
-
-    if command.is_empty() {
-        print_cursor_response(true, "allow", None, None);
-        return Ok(0);
-    }
-
-    match check_command_for_hook(&command) {
-        HookCheckResult::Allow => {
-            print_cursor_response(true, "allow", None, None);
-        }
-        HookCheckResult::BlockMeta(reason) => {
-            eprintln!("omamori cursor-hook: BLOCKED ({reason})");
-            print_cursor_response(
-                false,
-                "deny",
-                Some(&format!("omamori hook: {reason}")),
-                Some(&format!(
-                    "This command was blocked by omamori: {reason}. Use a safer alternative."
-                )),
-            );
-        }
-        HookCheckResult::BlockRule {
-            message,
-            unwrap_chain,
-            ..
-        } => {
-            let chain_str = unwrap_chain
-                .as_deref()
-                .map(|c| format!(" ({c})"))
-                .unwrap_or_default();
-            eprintln!("omamori cursor-hook: BLOCKED ({message}{chain_str})");
-            print_cursor_response(
-                false,
-                "deny",
-                Some(&format!("omamori hook: blocked — {message}{chain_str}")),
-                Some("This command was blocked by omamori safety guard. Use a safer alternative."),
-            );
-        }
-        HookCheckResult::BlockStructural(message) => {
-            eprintln!("omamori cursor-hook: BLOCKED ({message})");
-            print_cursor_response(
-                false,
-                "deny",
-                Some(&message),
-                Some("This command was blocked by omamori safety guard. Use a safer alternative."),
-            );
-        }
-    }
-
-    Ok(0)
-}
-
-// ---------------------------------------------------------------------------
-// Shared hook check logic (used by both hook-check and cursor-hook)
-// ---------------------------------------------------------------------------
-
-/// Result of checking a command string through the hook pipeline.
-enum HookCheckResult {
-    /// Command is allowed.
-    Allow,
-    /// Command is blocked by a meta-pattern (string-level).
-    BlockMeta(&'static str),
-    /// Command is blocked by the unwrap stack (token-level rule match).
-    BlockRule {
-        rule_name: String,
-        message: String,
-        unwrap_chain: Option<String>,
-    },
-    /// Command is blocked by the unwrap stack (structural block: pipe-to-shell, etc.).
-    BlockStructural(String),
-}
-
-/// Two-phase hook check:
-/// Phase 1: Meta-pattern string-level check (env var unset, config tamper, /bin/rm, etc.)
-/// Phase 2: Token-level unwrap stack → rule matching
-fn check_command_for_hook(command: &str) -> HookCheckResult {
-    // Phase 1: Meta-patterns (string-level, intentionally broad)
-    for (pattern, reason) in installer::blocked_command_patterns() {
-        if command.contains(pattern) {
-            return HookCheckResult::BlockMeta(reason);
-        }
-    }
-
-    // Phase 2: Unwrap stack → rule matching
-    let parse_result = unwrap::parse_command_string(command);
-
-    match parse_result {
-        unwrap::ParseResult::Block(reason) => HookCheckResult::BlockStructural(format!(
-            "omamori hook: blocked — {}",
-            reason.message()
-        )),
-        unwrap::ParseResult::Commands(invocations) => {
-            // Load config to get rules
-            let load_result = match load_config(None) {
-                Ok(r) => r,
-                Err(_) => {
-                    // Config load failure → use default rules (fail-safe, not fail-open)
-                    ConfigLoadResult {
-                        config: config::Config::default(),
-                        warnings: vec![],
-                    }
-                }
-            };
-
-            for inv in &invocations {
-                if let Some(rule) = match_rule(&load_result.config.rules, inv) {
-                    let chain_desc = format_unwrap_chain(command, inv);
-                    let msg = rule
-                        .message
-                        .clone()
-                        .unwrap_or_else(|| format!("matched rule: {}", rule.name));
-                    return HookCheckResult::BlockRule {
-                        rule_name: rule.name.clone(),
-                        message: msg,
-                        unwrap_chain: chain_desc,
-                    };
-                }
-            }
-
-            HookCheckResult::Allow
-        }
-    }
-}
-
-/// Format the unwrap chain for display: "rm -rf / (via bash -c)"
-fn format_unwrap_chain(original: &str, invocation: &CommandInvocation) -> Option<String> {
-    // If the original command doesn't start with the matched program,
-    // there was unwrapping involved — show the wrapper context
-    let trimmed = original.trim();
-    if !trimmed.starts_with(&invocation.program) {
-        // Extract the first word of the original as the outermost wrapper
-        let outer = trimmed.split_whitespace().next().unwrap_or("");
-        let outer_base = outer.rsplit('/').next().unwrap_or(outer);
-        // Check if bash -c style
-        if trimmed.contains("-c") {
-            Some(format!("via {} -c", outer_base))
-        } else {
-            Some(format!("via {}", outer_base))
-        }
-    } else {
-        None
-    }
-}
-
-// ---------------------------------------------------------------------------
-// hook-check subcommand (Claude Code PreToolUse thin wrapper target)
-// ---------------------------------------------------------------------------
-
-/// `omamori hook-check [--provider NAME]`
-/// Reads PreToolUse JSON from stdin, classifies via `HookInput`, then evaluates.
-/// Exit 0 = allow, exit 2 = block.
-fn run_hook_check(args: &[OsString]) -> Result<i32, AppError> {
-    use std::io::Read;
-
-    let provider = parse_provider_flag(args);
-    let verbose = std::env::var("OMAMORI_VERBOSE").is_ok();
-
-    let mut input = String::new();
-    std::io::stdin().read_to_string(&mut input)?;
-
-    match extract_hook_input(&input) {
-        HookInput::MalformedJson => {
-            eprintln!("omamori hook: blocked — hook input is not valid JSON");
-            eprintln!("  The command was denied because omamori cannot verify its safety.");
-            eprintln!(
-                "  This may happen after an AI tool update. Try: upgrade omamori, or report at https://github.com/yottayoshida/omamori/issues"
-            );
-            if verbose {
-                eprintln!("  provider: {provider}");
-                eprintln!(
-                    "  raw input (first 200 chars): {}",
-                    truncate_for_log(&input, 200)
-                );
-            }
-            Ok(2)
-        }
-        HookInput::MalformedMissingField => {
-            eprintln!("omamori hook: blocked — required fields missing from hook input");
-            eprintln!("  The command was denied because omamori cannot verify its safety.");
-            eprintln!("  Expected: tool_input.command or tool_input.file_path");
-            if verbose {
-                eprintln!("  provider: {provider}");
-                eprintln!(
-                    "  raw input (first 200 chars): {}",
-                    truncate_for_log(&input, 200)
-                );
-            }
-            Ok(2)
-        }
-        HookInput::UnknownTool(tool_name) => {
-            print_hook_check_allow_response(&format!(
-                "omamori: unrecognized tool '{tool_name}' — allowed for forward compatibility"
-            ));
-            Ok(0)
-        }
-        HookInput::FileOp { tool, path } => {
-            if let Some(reason) = is_protected_file_path(&path) {
-                eprintln!("omamori hook: blocked {tool} to protected file — {reason}");
-                eprintln!("  AI agents cannot modify omamori configuration or security files.");
-                eprintln!(
-                    "  To edit config: use `omamori config` CLI or edit the file directly in your terminal."
-                );
-                if verbose {
-                    eprintln!("  provider: {provider}");
-                    eprintln!("  tool: {tool}");
-                    eprintln!("  path: {path}");
-                }
-                Ok(2)
-            } else {
-                print_hook_check_allow_response(&format!(
-                    "omamori: {tool} to non-protected path — allowed"
-                ));
-                Ok(0)
-            }
-        }
-        HookInput::Command(command) => {
-            if command.is_empty() {
-                print_hook_check_allow_response("omamori: empty command");
-                return Ok(0);
-            }
-            run_hook_check_command(&command, &provider, verbose)
-        }
-    }
-}
-
-/// Evaluate a shell command through the two-phase hook check pipeline.
-/// Extracted from `run_hook_check` to keep the match arms concise.
-fn run_hook_check_command(command: &str, provider: &str, verbose: bool) -> Result<i32, AppError> {
-    match check_command_for_hook(command) {
-        HookCheckResult::Allow => {
-            print_hook_check_allow_response("omamori: no dangerous pattern detected");
-            Ok(0)
-        }
-        HookCheckResult::BlockMeta(reason) => {
-            eprintln!("omamori hook: blocked — {reason}");
-            if verbose {
-                eprintln!("  provider: {provider}");
-                eprintln!("  layer: meta-pattern (string-level)");
-            }
-            Ok(2)
-        }
-        HookCheckResult::BlockRule {
-            rule_name,
-            message,
-            unwrap_chain,
-        } => {
-            let chain_str = unwrap_chain
-                .as_deref()
-                .map(|c| format!(" ({c})"))
-                .unwrap_or_default();
-            eprintln!("omamori hook: blocked — {message}{chain_str}");
-            if verbose {
-                eprintln!("  provider: {provider}");
-                eprintln!("  rule: {rule_name}");
-                eprintln!("  layer: unwrap-stack (token-level)");
-            }
-            eprintln!(
-                "  hint: if intentional, run the command directly in your terminal (not via AI agent)"
-            );
-            Ok(2)
-        }
-        HookCheckResult::BlockStructural(message) => {
-            eprintln!("{message}");
-            if verbose {
-                eprintln!("  provider: {provider}");
-                eprintln!("  layer: unwrap-stack (structural)");
-            }
-            eprintln!(
-                "  hint: if intentional, run the command directly in your terminal (not via AI agent)"
-            );
-            Ok(2)
-        }
-    }
-}
-
-/// Truncate a string for log output, avoiding panic on multi-byte boundaries.
-fn truncate_for_log(s: &str, max_chars: usize) -> &str {
-    match s.char_indices().nth(max_chars) {
-        Some((idx, _)) => &s[..idx],
-        None => s,
-    }
-}
-
-// ---------------------------------------------------------------------------
-// File path protection for Edit/Write/MultiEdit (#110)
-// ---------------------------------------------------------------------------
-
-/// Patterns that identify omamori's own files and external hook registrations.
-/// Substring match against the canonicalized (or lexically normalized) path.
-/// Intentionally broad: "audit-secret" matches "audit-secret.1.retired" too.
-const PROTECTED_FILE_PATTERNS: &[(&str, &str)] = &[
-    ("omamori/config.toml", "omamori config"),
-    (".integrity.json", "integrity baseline"),
-    ("audit-secret", "audit HMAC secret"),
-    ("audit.jsonl", "audit log"),
-    (".local/share/omamori", "omamori data directory"),
-    ("claude-pretooluse.sh", "omamori hook script"),
-    ("codex-pretooluse.sh", "omamori Codex hook script"),
-    (".codex/hooks.json", "Codex hooks config"),
-    (".codex/config.toml", "Codex config"),
-    // Claude Code hook registration — AI removing hooks = full bypass (#110 T3)
-    (
-        ".claude/settings.json",
-        "Claude Code settings (contains hook config)",
-    ),
-];
-
-/// Check whether a file path targets a protected omamori file.
-///
-/// Resolution strategy (defense-in-depth per Codex review):
-///   1. Try `canonicalize()` to resolve symlinks for existing paths
-///   2. For non-existent paths, canonicalize the parent directory
-///   3. Fall back to lexical normalization (`context::normalize_path`)
-///   4. If canonicalize fails AND lexical path matches → fail-close (block)
-fn is_protected_file_path(path: &str) -> Option<&'static str> {
-    let lexical = context::normalize_path(path);
-
-    // Try full canonicalize first (resolves symlinks for existing paths)
-    let candidates: Vec<std::path::PathBuf> = match std::fs::canonicalize(&lexical) {
-        Ok(canonical) => vec![canonical],
-        Err(_) => {
-            // Full path doesn't exist — try canonicalizing the parent directory.
-            // This catches symlinked parents: /tmp/alias/newfile where /tmp/alias → protected dir
-            lexical
-                .parent()
-                .and_then(|p| std::fs::canonicalize(p).ok())
-                .and_then(|cp| lexical.file_name().map(|f| cp.join(f)))
-                .into_iter()
-                .collect()
-        }
-    };
-
-    // Check all resolved candidates + the lexical path itself
-    let lexical_str = lexical.to_string_lossy();
-    for &(pattern, reason) in PROTECTED_FILE_PATTERNS {
-        // Lexical match always checked (fail-close: even if canonicalize failed)
-        if lexical_str.contains(pattern) {
-            return Some(reason);
-        }
-        // Canonical/parent-resolved matches
-        for candidate in &candidates {
-            if candidate.to_string_lossy().contains(pattern) {
-                return Some(reason);
-            }
-        }
-    }
-    None
-}
-
-// ---------------------------------------------------------------------------
-// HookInput: typed representation of PreToolUse hook stdin
-// ---------------------------------------------------------------------------
-
-/// Parsed hook input from AI tool platforms (Claude Code, Codex CLI, etc.).
-///
-/// Fail-close design: anything that cannot be positively identified as a known
-/// tool invocation is treated as malformed and blocked.
-#[derive(Debug)]
-enum HookInput {
-    /// `tool_input.command` present — shell command to evaluate.
-    /// Covers tool_name = "Bash", "Shell", or any tool that carries a command field.
-    Command(String),
-
-    /// `tool_input.file_path` present — file operation (Edit/Write/MultiEdit).
-    /// Protected-path checking via `is_protected_file_path()`.
-    FileOp { tool: String, path: String },
-
-    /// Valid JSON with a `tool_name` but neither `command` nor `file_path`.
-    /// Allow for forward compatibility with future platform tool types.
-    UnknownTool(String),
-
-    /// JSON parse failed entirely (invalid syntax, non-UTF8, etc.).
-    MalformedJson,
-
-    /// JSON parsed but required structure is missing (no `tool_input`, etc.).
-    MalformedMissingField,
-}
-
-/// Parse PreToolUse hook stdin into a typed `HookInput`.
-///
-/// Classification priority:
-///   1. JSON parse failure → `MalformedJson`
-///   2. `tool_input.command` (string) → `Command`
-///   3. top-level `command` (Cursor compat) → `Command`
-///   4. `tool_input.file_path` (string) → `FileOp`
-///   5. `tool_name` present → `UnknownTool`
-///   6. nothing recognizable → `MalformedMissingField`
-fn extract_hook_input(input: &str) -> HookInput {
-    let v = match serde_json::from_str::<serde_json::Value>(input) {
-        Ok(v) => v,
-        Err(_) => return HookInput::MalformedJson,
-    };
-
-    let tool_input = v.get("tool_input");
-
-    // Priority 1: tool_input.command (Claude Code / Codex CLI format)
-    if let Some(cmd_val) = tool_input.and_then(|ti| ti.get("command")) {
-        // Key exists — value MUST be a string. null/number/array = malformed.
-        return match cmd_val.as_str() {
-            Some(cmd) => HookInput::Command(cmd.to_string()),
-            None => HookInput::MalformedMissingField,
-        };
-    }
-
-    // Priority 2: top-level command (Cursor beforeShellExecution compat)
-    if let Some(cmd_val) = v.get("command") {
-        return match cmd_val.as_str() {
-            Some(cmd) => HookInput::Command(cmd.to_string()),
-            None => HookInput::MalformedMissingField,
-        };
-    }
-
-    // Priority 3: tool_input.file_path (Edit/Write/MultiEdit)
-    if let Some(path_val) = tool_input.and_then(|ti| ti.get("file_path")) {
-        return match path_val.as_str() {
-            Some(path) => {
-                let tool = v
-                    .get("tool_name")
-                    .and_then(|t| t.as_str())
-                    .unwrap_or("unknown")
-                    .to_string();
-                HookInput::FileOp {
-                    tool,
-                    path: path.to_string(),
-                }
-            }
-            None => HookInput::MalformedMissingField,
-        };
-    }
-
-    // Priority 4: tool_input exists but has no command/file_path
-    if let Some(ti) = tool_input {
-        // Empty tool_input or non-object = malformed (e.g. {"tool_name":"Bash","tool_input":{}})
-        if ti.as_object().is_none_or(|obj| obj.is_empty()) {
-            return HookInput::MalformedMissingField;
-        }
-        // Non-empty tool_input with unrecognized fields = future tool type
-        if let Some(name) = v.get("tool_name").and_then(|t| t.as_str()) {
-            return HookInput::UnknownTool(name.to_string());
-        }
-        return HookInput::MalformedMissingField;
-    }
-
-    // Priority 5: no tool_input at all — tool_name alone = future tool type
-    if let Some(name) = v.get("tool_name").and_then(|t| t.as_str()) {
-        return HookInput::UnknownTool(name.to_string());
-    }
-
-    // Nothing recognizable
-    HookInput::MalformedMissingField
-}
-
-/// Parse --provider flag from args. Defaults to "unknown".
-fn parse_provider_flag(args: &[OsString]) -> String {
-    for (i, arg) in args.iter().enumerate() {
-        if arg.to_str() == Some("--provider")
-            && let Some(val) = args.get(i + 1)
-        {
-            return val.to_string_lossy().to_string();
-        }
-    }
-    "unknown".to_string()
-}
-
-/// Print hookSpecificOutput JSON for Claude Code PreToolUse hook (Auto mode compatibility).
-/// Only used on ALLOW path — BLOCK uses exit code 2 (stdout ignored by Claude Code).
-fn print_hook_check_allow_response(reason: &str) {
-    let response = serde_json::json!({
-        "hookSpecificOutput": {
-            "hookEventName": "PreToolUse",
-            "permissionDecision": "allow",
-            "permissionDecisionReason": reason,
-        }
-    });
-    // Fallback: if serialization somehow fails, emit a hardcoded JSON string (fail-safe)
-    println!(
-        "{}",
-        serde_json::to_string(&response).unwrap_or_else(|_| {
-            r#"{"hookSpecificOutput":{"hookEventName":"PreToolUse","permissionDecision":"allow","permissionDecisionReason":"omamori: fallback"}}"#.to_string()
-        })
-    );
-}
-
-fn print_cursor_response(
-    cont: bool,
-    permission: &str,
-    user_message: Option<&str>,
-    agent_message: Option<&str>,
-) {
-    let mut response = serde_json::json!({
-        "continue": cont,
-        "permission": permission,
-    });
-    if let Some(msg) = user_message {
-        response["userMessage"] = serde_json::json!(msg);
-    }
-    if let Some(msg) = agent_message {
-        response["agentMessage"] = serde_json::json!(msg);
-    }
-    // stdout is JSON only — never print anything else here
-    println!(
-        "{}",
-        serde_json::to_string(&response)
-            .unwrap_or_else(|_| { r#"{"continue":false,"permission":"deny"}"#.to_string() })
-    );
-}
-
 fn run_config_command(args: &[OsString]) -> Result<i32, AppError> {
     match args.get(2).and_then(|item| item.to_str()) {
         Some("list") => run_config_list(),
@@ -1717,23 +814,6 @@ fn run_config_command(args: &[OsString]) -> Result<i32, AppError> {
             usage_text()
         ))),
     }
-}
-
-/// Guard against AI agents modifying omamori's own configuration.
-/// Blocks when any AI detector env var is present (exact match via evaluate_detectors).
-fn guard_ai_config_modification(operation: &str) -> Result<(), AppError> {
-    let detectors = config::default_detectors();
-    let env_pairs: Vec<(String, String)> = std::env::vars().collect();
-    let detection = evaluate_detectors(&detectors, &env_pairs);
-    if detection.protected {
-        return Err(AppError::Config(format!(
-            "{operation} blocked — AI agent environment detected ({}).\n  \
-             Protection rules cannot be modified by AI tools.\n  \
-             To modify, run this command directly in your terminal (not via AI).",
-            detection.matched_detectors.join(", ")
-        )));
-    }
-    Ok(())
 }
 
 fn validate_rule_name(name: &str) -> Result<(), AppError> {
@@ -2173,18 +1253,6 @@ pub struct PolicyTestResult {
     pub details: String,
 }
 
-/// Fuzz entry point: parse hook input without exposing internal types.
-/// Panics are the signal — if this doesn't panic, the input is handled safely.
-pub fn fuzz_extract_hook_input(input: &str) {
-    let _ = extract_hook_input(input);
-}
-
-/// Fuzz entry point: check a command string through the hook pipeline.
-/// Exercises meta-patterns + unwrap stack + rule matching.
-pub fn fuzz_check_command_for_hook(command: &str) {
-    let _ = check_command_for_hook(command);
-}
-
 pub fn run_policy_tests(load_result: &ConfigLoadResult) -> Vec<PolicyTestResult> {
     let config = &load_result.config;
     let claude_env = vec![("CLAUDECODE".to_string(), "1".to_string())];
@@ -2386,161 +1454,6 @@ mod tests {
 
     // --- G-02: ensure_hooks_current_at ---
 
-    fn setup_hooks_dir(base_dir: &Path) -> PathBuf {
-        let hooks_dir = base_dir.join("hooks");
-        std::fs::create_dir_all(&hooks_dir).unwrap();
-        hooks_dir
-    }
-
-    #[test]
-    fn hooks_current_old_version_triggers_regen() {
-        let dir = std::env::temp_dir().join(format!("omamori-hooks-g02-1-{}", std::process::id()));
-        let _ = std::fs::remove_dir_all(&dir);
-        let hooks_dir = setup_hooks_dir(&dir);
-
-        // Write hook with old version
-        let old_hook = "#!/bin/sh\n# omamori hook v0.0.1\nset -eu\nexit 0\n";
-        std::fs::write(hooks_dir.join("claude-pretooluse.sh"), old_hook).unwrap();
-
-        let result = ensure_hooks_current_at(&dir);
-        assert!(result, "should regenerate hooks for old version");
-
-        // Verify the regenerated hook has the current version
-        let content = std::fs::read_to_string(hooks_dir.join("claude-pretooluse.sh")).unwrap();
-        assert_eq!(
-            installer::parse_hook_version(&content),
-            Some(env!("CARGO_PKG_VERSION"))
-        );
-
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn hooks_current_hash_mismatch_triggers_regen() {
-        let dir = std::env::temp_dir().join(format!("omamori-hooks-g02-2-{}", std::process::id()));
-        let _ = std::fs::remove_dir_all(&dir);
-        let hooks_dir = setup_hooks_dir(&dir);
-
-        // Write hook with correct version but tampered body (T2 attack)
-        let tampered = format!(
-            "#!/bin/sh\n# omamori hook v{}\nset -eu\nexit 0\n",
-            env!("CARGO_PKG_VERSION")
-        );
-        std::fs::write(hooks_dir.join("claude-pretooluse.sh"), tampered).unwrap();
-
-        let result = ensure_hooks_current_at(&dir);
-        assert!(result, "should regenerate hooks for hash mismatch (T2)");
-
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn hooks_current_correct_returns_false() {
-        let dir = std::env::temp_dir().join(format!("omamori-hooks-g02-3-{}", std::process::id()));
-        let _ = std::fs::remove_dir_all(&dir);
-        let hooks_dir = setup_hooks_dir(&dir);
-
-        // Write the exact expected hook content
-        let expected = installer::render_hook_script();
-        std::fs::write(hooks_dir.join("claude-pretooluse.sh"), expected).unwrap();
-
-        let result = ensure_hooks_current_at(&dir);
-        assert!(!result, "should return false when hooks are current");
-
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn hooks_current_missing_returns_false() {
-        let dir = std::env::temp_dir().join(format!("omamori-hooks-g02-4-{}", std::process::id()));
-        let _ = std::fs::remove_dir_all(&dir);
-        std::fs::create_dir_all(&dir).unwrap();
-
-        // No hooks file at all
-        let result = ensure_hooks_current_at(&dir);
-        assert!(!result, "should return false when no hooks file");
-
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn hooks_current_readonly_dir_regen_fails_returns_false() {
-        let dir = std::env::temp_dir().join(format!("omamori-hooks-g02-5-{}", std::process::id()));
-        let _ = std::fs::remove_dir_all(&dir);
-        let hooks_dir = setup_hooks_dir(&dir);
-
-        // Write hook with old version
-        let old_hook = "#!/bin/sh\n# omamori hook v0.0.1\nset -eu\nexit 0\n";
-        std::fs::write(hooks_dir.join("claude-pretooluse.sh"), old_hook).unwrap();
-
-        // Make hooks dir read-only so regeneration fails
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            std::fs::set_permissions(&hooks_dir, std::fs::Permissions::from_mode(0o555)).unwrap();
-        }
-
-        let result = ensure_hooks_current_at(&dir);
-        assert!(!result, "should return false when regen fails");
-
-        // Restore permissions for cleanup
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            std::fs::set_permissions(&hooks_dir, std::fs::Permissions::from_mode(0o755)).unwrap();
-        }
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    // --- G-03: should_block_for_sudo ---
-
-    // --- ADV-01: hooks symlink attack ---
-
-    #[test]
-    fn hooks_symlink_attack_triggers_regen() {
-        // ADV-01: If the hooks file is a symlink (attacker replaced it),
-        // ensure_hooks_current_at reads through it and detects the hash mismatch.
-        let dir = std::env::temp_dir().join(format!("omamori-adv01-{}", std::process::id()));
-        let _ = std::fs::remove_dir_all(&dir);
-        let hooks_dir = setup_hooks_dir(&dir);
-
-        #[cfg(unix)]
-        {
-            // Create a malicious script elsewhere
-            let malicious = dir.join("malicious.sh");
-            std::fs::write(
-                &malicious,
-                format!(
-                    "#!/bin/sh\n# omamori hook v{}\nexit 0\n",
-                    env!("CARGO_PKG_VERSION")
-                ),
-            )
-            .unwrap();
-
-            // Replace hook with symlink to malicious script
-            let hook_path = hooks_dir.join("claude-pretooluse.sh");
-            std::os::unix::fs::symlink(&malicious, &hook_path).unwrap();
-
-            // ensure_hooks_current_at should detect hash mismatch and regenerate
-            let result = ensure_hooks_current_at(&dir);
-            assert!(
-                result,
-                "symlink hook should trigger regeneration due to hash mismatch"
-            );
-
-            // After regen, the hook should have the correct hash
-            let content = std::fs::read_to_string(&hook_path).unwrap();
-            let expected = installer::render_hook_script();
-            assert_eq!(
-                installer::hook_content_hash(&content),
-                installer::hook_content_hash(&expected),
-                "regenerated hook should match expected content"
-            );
-        }
-
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
     // =====================================================================
     // Guardrail tests for v0.8.1 module split (#112)
     //
@@ -2548,346 +1461,74 @@ mod tests {
     // NOT directly cover.  They must pass before AND after each PR in the
     // module split sequence.
     // =====================================================================
+}
 
-    // --- GR-001: fail-close config fallback (T8 guardrail, DREAD 9.0) ---
-    //
-    // These tests isolate from ambient config by pointing XDG_CONFIG_HOME
-    // and HOME to an empty temp dir.  load_config(None) will fail to find
-    // config.toml and fall back to Config::default() — exactly the T8
-    // fail-close path we want to verify.
+// --- GR-005: mutate_config pipeline (T6 guardrail) ---
 
-    /// Redirect config discovery to an empty dir so load_config falls back
-    /// to Config::default().  Returns (old_xdg, old_home) for restore.
-    /// Redirect config discovery to an empty temp dir so `load_config(None)`
-    /// cannot find config.toml and falls back to `Config::default()`.
-    ///
-    /// # Safety
-    /// Env-var mutation is guarded by `#[serial_test::serial]` on every caller.
-    fn isolate_config() -> (Option<String>, Option<String>, std::path::PathBuf) {
-        let dir = std::env::temp_dir().join(format!("omamori-gr-iso-{}", std::process::id()));
-        let _ = std::fs::remove_dir_all(&dir);
-        std::fs::create_dir_all(&dir).unwrap();
-        let old_xdg = std::env::var("XDG_CONFIG_HOME").ok();
-        let old_home = std::env::var("HOME").ok();
-        // SAFETY: serialized by #[serial_test::serial] — no concurrent env reads.
-        unsafe {
-            std::env::set_var("XDG_CONFIG_HOME", dir.join("xdg"));
-            std::env::set_var("HOME", &dir);
-        }
-        (old_xdg, old_home, dir)
-    }
+#[test]
+fn mutate_config_rejects_invalid_mutation() {
+    let dir = std::env::temp_dir().join(format!("omamori-gr005-1-{}", std::process::id()));
+    let _ = std::fs::remove_dir_all(&dir);
+    std::fs::create_dir_all(&dir).unwrap();
+    let config_path = dir.join("config.toml");
+    std::fs::write(&config_path, "[rules]\n").unwrap();
 
-    fn restore_config(old_xdg: Option<String>, old_home: Option<String>, dir: std::path::PathBuf) {
-        // SAFETY: serialized by #[serial_test::serial] — no concurrent env reads.
-        unsafe {
-            match old_xdg {
-                Some(v) => std::env::set_var("XDG_CONFIG_HOME", v),
-                None => std::env::remove_var("XDG_CONFIG_HOME"),
-            }
-            match old_home {
-                Some(v) => std::env::set_var("HOME", v),
-                None => std::env::remove_var("HOME"),
-            }
-        }
-        let _ = std::fs::remove_dir_all(dir);
-    }
+    let original = std::fs::read_to_string(&config_path).unwrap();
 
-    #[test]
-    #[serial_test::serial]
-    fn check_command_for_hook_blocks_rm_rf_with_default_rules() {
-        let (old_xdg, old_home, dir) = isolate_config();
+    // Mutation that produces invalid TOML (unparseable by the failsafe toml crate)
+    let result = mutate_config(&config_path, |doc| {
+        // Insert raw string that is valid toml_edit but breaks toml's stricter parser
+        doc.insert("__broken", toml_edit::Item::None);
+        Ok(())
+    });
 
-        match check_command_for_hook("rm -rf /") {
-            HookCheckResult::BlockRule { rule_name, .. } => {
-                assert!(
-                    rule_name.contains("rm"),
-                    "expected rm-related rule, got: {rule_name}"
-                );
-            }
-            HookCheckResult::BlockMeta(_) | HookCheckResult::BlockStructural(_) => {
-                // Also acceptable — blocked by meta-pattern or structural check.
-            }
-            HookCheckResult::Allow => {
-                restore_config(old_xdg, old_home, dir);
-                panic!("SECURITY: rm -rf / was ALLOWED — fail-close fallback is broken");
-            }
-        }
-        restore_config(old_xdg, old_home, dir);
-    }
-
-    #[test]
-    #[serial_test::serial]
-    fn check_command_for_hook_allows_safe_command() {
-        let (old_xdg, old_home, dir) = isolate_config();
-
-        match check_command_for_hook("ls /tmp") {
-            HookCheckResult::Allow => {} // expected
-            other => {
-                restore_config(old_xdg, old_home, dir);
-                panic!(
-                    "expected Allow for 'ls /tmp', got: {}",
-                    match other {
-                        HookCheckResult::BlockMeta(r) => format!("BlockMeta({r})"),
-                        HookCheckResult::BlockRule { rule_name, .. } =>
-                            format!("BlockRule({rule_name})"),
-                        HookCheckResult::BlockStructural(r) => format!("BlockStructural({r})"),
-                        HookCheckResult::Allow => unreachable!(),
-                    }
-                );
-            }
-        }
-        restore_config(old_xdg, old_home, dir);
-    }
-
-    // --- GR-003: extract_hook_input 6-class unit tests (pipeline guardrail) ---
-
-    #[test]
-    fn extract_hook_input_command_from_tool_input() {
-        let input = r#"{"tool_name":"Bash","tool_input":{"command":"ls -la"}}"#;
-        match extract_hook_input(input) {
-            HookInput::Command(cmd) => assert_eq!(cmd, "ls -la"),
-            other => panic!("expected Command, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn extract_hook_input_command_from_top_level() {
-        // Cursor beforeShellExecution compat
-        let input = r#"{"command":"echo hello"}"#;
-        match extract_hook_input(input) {
-            HookInput::Command(cmd) => assert_eq!(cmd, "echo hello"),
-            other => panic!("expected Command, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn extract_hook_input_file_op() {
-        let input = r#"{"tool_name":"Edit","tool_input":{"file_path":"/tmp/x.rs"}}"#;
-        match extract_hook_input(input) {
-            HookInput::FileOp { tool, path } => {
-                assert_eq!(tool, "Edit");
-                assert_eq!(path, "/tmp/x.rs");
-            }
-            other => panic!("expected FileOp, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn extract_hook_input_unknown_tool() {
-        let input = r#"{"tool_name":"FutureTool","tool_input":{"query":"something"}}"#;
-        match extract_hook_input(input) {
-            HookInput::UnknownTool(name) => assert_eq!(name, "FutureTool"),
-            other => panic!("expected UnknownTool, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn extract_hook_input_malformed_json() {
-        match extract_hook_input("not json at all") {
-            HookInput::MalformedJson => {}
-            other => panic!("expected MalformedJson, got: {other:?}"),
-        }
-    }
-
-    #[test]
-    fn extract_hook_input_missing_field() {
-        // tool_input exists but is empty object → MalformedMissingField
-        let input = r#"{"tool_name":"Bash","tool_input":{}}"#;
-        match extract_hook_input(input) {
-            HookInput::MalformedMissingField => {}
-            other => panic!("expected MalformedMissingField, got: {other:?}"),
-        }
-    }
-
-    // --- GR-004: is_protected_file_path (PROTECTED_FILE_PATTERNS guardrail) ---
-
-    #[test]
-    fn protected_file_path_matches_config_toml() {
-        let result = is_protected_file_path("/home/user/.config/omamori/config.toml");
-        assert!(result.is_some(), "config.toml should be protected");
-    }
-
-    #[test]
-    fn protected_file_path_rejects_unrelated() {
-        let result = is_protected_file_path("/tmp/myfile.txt");
-        assert!(result.is_none(), "/tmp/myfile.txt should not be protected");
-    }
-
-    #[test]
-    fn protected_file_path_all_patterns_match() {
-        // Verify every pattern in PROTECTED_FILE_PATTERNS is reachable.
-        let test_paths = [
-            "/home/user/.config/omamori/config.toml",
-            "/home/user/.local/share/omamori/.integrity.json",
-            "/home/user/.local/share/omamori/audit-secret",
-            "/home/user/.local/share/omamori/audit.jsonl",
-            "/home/user/.local/share/omamori",
-            "/home/user/.local/share/omamori/hooks/claude-pretooluse.sh",
-            "/home/user/.local/share/omamori/hooks/codex-pretooluse.sh",
-            "/home/user/.codex/hooks.json",
-            "/home/user/.codex/config.toml",
-            "/home/user/.claude/settings.json",
-        ];
-        for path in &test_paths {
-            assert!(
-                is_protected_file_path(path).is_some(),
-                "PROTECTED_FILE_PATTERNS gap: {path} was not matched"
-            );
-        }
-    }
-
-    // --- GR-007: check_command_for_hook meta-pattern (hook pipeline guardrail) ---
-
-    #[test]
-    #[serial_test::serial]
-    fn check_command_for_hook_blocks_meta_pattern() {
-        let (old_xdg, old_home, dir) = isolate_config();
-
-        match check_command_for_hook("unset CLAUDECODE") {
-            HookCheckResult::BlockMeta(_) => {}
-            HookCheckResult::BlockRule { .. } | HookCheckResult::BlockStructural(_) => {}
-            HookCheckResult::Allow => {
-                restore_config(old_xdg, old_home, dir);
-                panic!("SECURITY: 'unset CLAUDECODE' was ALLOWED — meta-pattern is broken");
-            }
-        }
-        restore_config(old_xdg, old_home, dir);
-    }
-
-    #[test]
-    #[serial_test::serial]
-    fn check_command_for_hook_allows_echo() {
-        let (old_xdg, old_home, dir) = isolate_config();
-
-        match check_command_for_hook("echo hello world") {
-            HookCheckResult::Allow => {}
-            _ => {
-                restore_config(old_xdg, old_home, dir);
-                panic!("'echo hello world' should be allowed");
-            }
-        }
-        restore_config(old_xdg, old_home, dir);
-    }
-
-    // --- GR-005: mutate_config pipeline (T6 guardrail) ---
-
-    #[test]
-    fn mutate_config_rejects_invalid_mutation() {
-        let dir = std::env::temp_dir().join(format!("omamori-gr005-1-{}", std::process::id()));
-        let _ = std::fs::remove_dir_all(&dir);
-        std::fs::create_dir_all(&dir).unwrap();
-        let config_path = dir.join("config.toml");
-        std::fs::write(&config_path, "[rules]\n").unwrap();
-
-        let original = std::fs::read_to_string(&config_path).unwrap();
-
-        // Mutation that produces invalid TOML (unparseable by the failsafe toml crate)
-        let result = mutate_config(&config_path, |doc| {
-            // Insert raw string that is valid toml_edit but breaks toml's stricter parser
-            doc.insert("__broken", toml_edit::Item::None);
-            Ok(())
-        });
-
-        // Regardless of whether this particular mutation triggers the failsafe,
-        // the original file must not be corrupted.
-        let after = std::fs::read_to_string(&config_path).unwrap_or_default();
-        // If mutation succeeded, the file was validly updated (also acceptable).
-        // The key invariant: the file is never left in a corrupt state.
-        if result.is_err() {
-            assert_eq!(
-                after, original,
-                "config must not be corrupted on mutation error"
-            );
-        } else {
-            // Mutation succeeded — verify output is valid TOML
-            assert!(
-                toml::from_str::<toml::Value>(&after).is_ok(),
-                "mutate_config produced invalid TOML"
-            );
-        }
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn mutate_config_roundtrip_preserves_structure() {
-        let dir = std::env::temp_dir().join(format!("omamori-gr005-2-{}", std::process::id()));
-        let _ = std::fs::remove_dir_all(&dir);
-        std::fs::create_dir_all(&dir).unwrap();
-        let config_path = dir.join("config.toml");
-
-        // Write a minimal valid config
-        let initial = "[rules]\n[audit]\nenabled = true\n";
-        std::fs::write(&config_path, initial).unwrap();
-
-        // Apply a valid mutation
-        let result = mutate_config(&config_path, |doc| {
-            doc["audit"]["enabled"] = toml_edit::value(false);
-            Ok(())
-        });
-        assert!(result.is_ok(), "valid mutation should succeed: {result:?}");
-
-        let after = std::fs::read_to_string(&config_path).unwrap();
-        let parsed: toml::Value = toml::from_str(&after).expect("output must be valid TOML");
+    // Regardless of whether this particular mutation triggers the failsafe,
+    // the original file must not be corrupted.
+    let after = std::fs::read_to_string(&config_path).unwrap_or_default();
+    // If mutation succeeded, the file was validly updated (also acceptable).
+    // The key invariant: the file is never left in a corrupt state.
+    if result.is_err() {
         assert_eq!(
-            parsed
-                .get("audit")
-                .and_then(|a| a.get("enabled"))
-                .and_then(|v| v.as_bool()),
-            Some(false),
-            "mutation should have set audit.enabled = false"
+            after, original,
+            "config must not be corrupted on mutation error"
         );
-        let _ = std::fs::remove_dir_all(&dir);
+    } else {
+        // Mutation succeeded — verify output is valid TOML
+        assert!(
+            toml::from_str::<toml::Value>(&after).is_ok(),
+            "mutate_config produced invalid TOML"
+        );
     }
+    let _ = std::fs::remove_dir_all(&dir);
+}
 
-    // --- GR-008: try_audit_append strict/non-strict (audit safety guardrail) ---
+#[test]
+fn mutate_config_roundtrip_preserves_structure() {
+    let dir = std::env::temp_dir().join(format!("omamori-gr005-2-{}", std::process::id()));
+    let _ = std::fs::remove_dir_all(&dir);
+    std::fs::create_dir_all(&dir).unwrap();
+    let config_path = dir.join("config.toml");
 
-    #[test]
-    fn try_audit_append_strict_returns_exit_code_on_failure() {
-        // Create a logger pointing at a non-existent directory → append will fail
-        let logger = audit::AuditLogger::from_config(&audit::AuditConfig {
-            enabled: true,
-            path: Some(std::path::PathBuf::from("/nonexistent/dir/audit.jsonl")),
-            retention_days: 0,
-            strict: true,
-        });
-        if let Some(logger) = logger {
-            let event = logger.create_event(
-                &rules::CommandInvocation::new("test".to_string(), vec![]),
-                None,
-                &[],
-                &actions::ActionOutcome::PassedThrough { exit_code: 0 },
-            );
-            let result = try_audit_append(&logger, event, true);
-            assert_eq!(
-                result,
-                Some(1),
-                "strict mode should return Some(1) on append failure"
-            );
-        }
-        // If logger construction itself fails (no secret), that's also acceptable —
-        // the point is that strict+failure → exit code.
-    }
+    // Write a minimal valid config
+    let initial = "[rules]\n[audit]\nenabled = true\n";
+    std::fs::write(&config_path, initial).unwrap();
 
-    #[test]
-    fn try_audit_append_non_strict_returns_none_on_failure() {
-        let logger = audit::AuditLogger::from_config(&audit::AuditConfig {
-            enabled: true,
-            path: Some(std::path::PathBuf::from("/nonexistent/dir/audit.jsonl")),
-            retention_days: 0,
-            strict: false,
-        });
-        if let Some(logger) = logger {
-            let event = logger.create_event(
-                &rules::CommandInvocation::new("test".to_string(), vec![]),
-                None,
-                &[],
-                &actions::ActionOutcome::PassedThrough { exit_code: 0 },
-            );
-            let result = try_audit_append(&logger, event, false);
-            assert_eq!(
-                result, None,
-                "non-strict mode should return None on append failure"
-            );
-        }
-    }
+    // Apply a valid mutation
+    let result = mutate_config(&config_path, |doc| {
+        doc["audit"]["enabled"] = toml_edit::value(false);
+        Ok(())
+    });
+    assert!(result.is_ok(), "valid mutation should succeed: {result:?}");
+
+    let after = std::fs::read_to_string(&config_path).unwrap();
+    let parsed: toml::Value = toml::from_str(&after).expect("output must be valid TOML");
+    assert_eq!(
+        parsed
+            .get("audit")
+            .and_then(|a| a.get("enabled"))
+            .and_then(|v| v.as_bool()),
+        Some(false),
+        "mutation should have set audit.enabled = false"
+    );
+    let _ = std::fs::remove_dir_all(&dir);
 }
